# [tests] Speed up test_unlikelihood