**Patch description**
Slow tests be slow. Drop a few redundant tests (identical models that only differed in a hyperparameter from the paper), delete tests that were being skipped, and significantly reduce computation in the others.

**Testing steps**

Before:
```
136.48s call     tests/nightly/gpu/test_unlikelihood.py::TestUnlikelihood::test_train_model_repeat_ul
136.43s call     tests/nightly/gpu/test_unlikelihood.py::TestUnlikelihood::test_train_model_vocab_ul
107.48s call     tests/nightly/gpu/test_unlikelihood.py::TestUnlikelihood::test_vocab_alpha1e3
100.23s call     tests/nightly/gpu/test_unlikelihood.py::TestUnlikelihood::test_vocab_alpha1e2
61.46s call     tests/nightly/gpu/test_unlikelihood.py::TestUnlikelihood::test_repeat_wiki_labelonly
59.10s call     tests/nightly/gpu/test_unlikelihood.py::TestUnlikelihood::test_repeat_wiki_contextandlabel
57.25s call     tests/nightly/gpu/test_unlikelihood.py::TestUnlikelihood::test_repeat_wiki_contextonly
39.91s call     tests/nightly/gpu/test_unlikelihood.py::TestUnlikelihood::test_vocab_alpha1e0
38.64s call     tests/nightly/gpu/test_unlikelihood.py::TestUnlikelihood::test_vocab_alpha1e1
27.01s call     tests/nightly/gpu/test_unlikelihood.py::TestUnlikelihood::test_repeat_convai_contextandlabel
======= 12 passed, 3 skipped, 113 warnings in 811.53s (0:13:31) =======
```

After:
```
$ pytest tests/nightly/gpu/test_unlikelihood.py
tests/nightly/gpu/test_unlikelihood.py .......... [100%]

56.96s call     tests/nightly/gpu/test_unlikelihood.py::TestUnlikelihood::test_train_model_vocab_ul
32.12s call     tests/nightly/gpu/test_unlikelihood.py::TestUnlikelihood::test_repeat_wiki_contextandlabel
30.61s call     tests/nightly/gpu/test_unlikelihood.py::TestUnlikelihood::test_repeat_wiki_contextonly
28.69s call     tests/nightly/gpu/test_unlikelihood.py::TestUnlikelihood::test_repeat_wiki_labelonly
12.26s call     tests/nightly/gpu/test_unlikelihood.py::TestUnlikelihood::test_vocab_alpha1e0
11.91s call     tests/nightly/gpu/test_unlikelihood.py::TestUnlikelihood::test_train_model_repeat_ul
11.86s call     tests/nightly/gpu/test_unlikelihood.py::TestUnlikelihood::test_vocab_alpha1e1
11.39s call     tests/nightly/gpu/test_unlikelihood.py::TestUnlikelihood::test_repeat_convai_contextandlabel
7.30s call     tests/nightly/gpu/test_unlikelihood.py::TestUnlikelihood::test_repeat_convai_contextonly
7.14s call     tests/nightly/gpu/test_unlikelihood.py::TestUnlikelihood::test_repeat_convai_labelonly
======= 10 passed, 113 warnings in 213.02s (0:03:33) =======
```